+++ include/globals.h
@@ -6,6 +6,6 @@
 
 extern file_handle_t logger;
 extern editor_t      editor;
-extern size_t        line_pad;
+extern unsigned int  line_pad;
 
 #endif /* GLOBALS_H */
